third_party/WebKit/LayoutTests/animations/font-size-using-ems.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="box"></div>
<style>
    body {
        font-size: 30px;
    }
    #box {
        animation-duration: 3s;
        animation-timing-function: linear;
        animation-delay: -1s;
        animation-play-state: paused;
        animation-name: anim;
    }
    @keyframes anim {
        from { font-size: 1em; }
        to   { font-size: 2em; }
    }
</style>
<script>
test(() => {
    assert_equals(getComputedStyle(box).fontSize, '40px');
}, 'font-size animation using em units gets the default font size');
</script>
